updated some names, organized into folders, and started modularizing
