+++ bin/index.js
