+++ commands/templates.js
