+++ package.json
@@ -4,5 +4,15 @@
     "date-fns": "^3.3.1",
     "express": "^4.18.2",
     "uuid": "^9.0.1"
-  }
+  },
+  "name": "s3-midtermsprint-fs",
+  "description": "",
+  "version": "1.0.0",
+  "main": "app.js",
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
 }
